On peut resize
Tout est dans le titre. Petite problème avec la 2D.

+++ src/parametres/Parametres.java
@@ -2,13 +2,11 @@ package parametres;
 
 public interface Parametres {
 
-	public float largeur = 960;
-	public float hauteur = 600;
+	public int largeur = 960;
+	public int hauteur = 600;
 	
 	float fov = 50;
 	float near = 0.01f;
 	float far = 45f;
-	
-	
 	
 }
